Revamp algorithms and enable algorithms test
