+++ include/mpp/algo/lu.hpp
@@ -22,7 +22,7 @@
 #include <mpp/detail/util/algo_impl.hpp>
 #include <mpp/detail/util/cpo_base.hpp>
 #include <mpp/detail/util/util.hpp>
-#include <mpp/mat/matfwd.hpp>
+#include <mpp/mat.hpp>
 
 #include <cassert>
 #include <cstddef>
@@ -31,45 +31,36 @@ namespace mpp
 {
 	namespace detail
 	{
-		template<typename To, typename To2, typename Mat>
-		auto lu_impl(const Mat& obj) -> std::pair<To, To2> // @TODO: ISSUE #20
+		template<typename To, typename Mat>
+		auto lu_impl(const Mat& obj) -> std::pair<To, To> // @TODO: ISSUE #20
 		{
 			assert(obj.rows() == obj.cols());
 
-			using lu_buffer_t = typename mat_rebind_to_t<Mat, fp_t>::buffer_type;
+			using val_t = typename To::value_type;
+			using buf_t = typename To::buffer_type;
 
 			const auto rows = obj.rows();
 			const auto cols = obj.cols();
 
-			auto l = lu_buffer_t{};
-			auto u = lu_buffer_t{};
+			auto l = buf_t{};
+			auto u = copy_into_new_buf<buf_t>(obj, rows, cols);
 
-			// @TODO: Should do a direct buffer copy initialization instead
-			resize_buf_if_vec(u, rows, cols, fp_t{});
-			std::ranges::copy(obj, u.begin());
+			init_identity_buf(l, rows, cols, val_t{}, val_t{ 1 });
 
-			init_identity_buf(l, rows, cols, fp_t{}, fp_t{ 1 });
+			lu_algo<buf_t, true, false>(rows, cols, l, u);
 
-			lu_impl<fp_t, true, false>(rows, cols, l, u);
-
-			return { To{ rows, cols, std::move(l) }, To2{ rows, cols, std::move(u) } };
+			return { To{ rows, cols, std::move(l) }, To{ rows, cols, std::move(u) } };
 		}
 	} // namespace detail
 
 	struct lu_t : public detail::cpo_base<lu_t>
 	{
-		template<typename Val,
-			std::size_t Rows,
-			std::size_t Cols,
-			typename Alloc,
-			typename To  = mat<Val, Rows, Cols, Alloc>,
-			typename To2 = mat<Val, Rows, Cols, Alloc>>
-		requires(detail::is_mat<To>::value&& detail::is_mat<To2>::value) friend inline auto tag_invoke(lu_t,
-			const mat<Val, Rows, Cols, Alloc>& obj,
-			std::type_identity<To>  = {},
-			std::type_identity<To2> = {}) -> std::pair<To, To2> // @TODO: ISSUE #20
+		template<typename T, typename Buf, typename To = mat<T, Buf>>
+		requires(detail::is_mat<To>::value) friend inline auto tag_invoke(lu_t,
+			const mat<T, Buf>& obj,
+			std::type_identity<To> = {}) -> std::pair<To, To> // @TODO: ISSUE #20
 		{
-			return detail::lu_impl<To, To2>(obj);
+			return detail::lu_impl<To>(obj);
 		}
 	};
