+++ include/mpp/algo/trps.hpp
@@ -21,7 +21,7 @@
 
 #include <mpp/detail/util/cpo_base.hpp>
 #include <mpp/detail/util/util.hpp>
-#include <mpp/mat/matfwd.hpp>
+#include <mpp/mat.hpp>
 
 #include <cstddef>
 #include <type_traits>
@@ -37,10 +37,8 @@ namespace mpp
 			const auto cols = obj.cols();
 			const auto data = obj.data();
 
-			using trps_buf_t = typename To::buffer_type;
-
-			auto trps_buf = trps_buf_t{};
-			resize_buf_if_vec(trps_buf, cols, rows, typename To::value_type{});
+			auto buf = typename To::buffer_type{};
+			resize_buf_if_dyn(buf, cols, rows, typename To::value_type{});
 
 			for (auto col = std::size_t{}; col < cols; ++col)
 			{
@@ -49,24 +47,20 @@ namespace mpp
 					auto idx      = detail::idx_1d(cols, row, col);
 					auto trps_idx = detail::idx_1d(rows, col, row);
 
-					trps_buf[trps_idx] = data[idx];
+					buf[trps_idx] = data[idx];
 				}
 			}
 
-			return To{ cols, rows, std::move(trps_buf) };
+			return To{ cols, rows, std::move(buf) };
 		}
 	} // namespace detail
 
 	struct trps_t : public detail::cpo_base<trps_t>
 	{
-		template<typename Val,
-			std::size_t Rows,
-			std::size_t Cols,
-			typename Alloc,
-			typename To = mat<Val, Cols, Rows, Alloc>>
-		requires(detail::is_mat<To>::value) [[nodiscard]] friend inline auto tag_invoke(trps_t,
-			const mat<Val, Rows, Cols, Alloc>& obj,
-			std::type_identity<To> = {}) -> To // @TODO: ISSUE #20
+		template<typename T, typename Buf, typename To = mat<T, Buf>>
+		requires(detail::is_mat<To>::value)
+			[[nodiscard]] friend inline auto tag_invoke(trps_t, const mat<T, Buf>& obj, std::type_identity<To> = {})
+				-> To // @TODO: ISSUE #20
 		{
 			return detail::trps_impl<To>(obj);
 		}
