+++ include/mpp/util/cmp.hpp
@@ -21,7 +21,7 @@
 
 #include <mpp/detail/util/cpo_base.hpp>
 #include <mpp/detail/util/util.hpp>
-#include <mpp/mat/matfwd.hpp>
+#include <mpp/mat.hpp>
 
 #include <algorithm>
 #include <cmath>
