+++ include/mpp/util/print.hpp
@@ -21,7 +21,7 @@
 
 #include <mpp/detail/util/cpo_base.hpp>
 #include <mpp/detail/util/print_impl.hpp>
-#include <mpp/mat/matfwd.hpp>
+#include <mpp/mat.hpp>
 
 #include <cstddef>
 #include <iosfwd>
